decommission One Wiley Living Style Guide
The One Wiley Living Style Guide was based on an early version of the Wiley corporate brand guidelines. It was also built to serve the old Wiley Online Library before the migration to Atypon's platform, Literatum. Since the move to Literatum, a new style guide is in place that serves Wiley Online Library and connected products.